+++ src/schorr_euchner.c
@@ -6,7 +6,7 @@
 #include "vector.h"
 #include "vector2d.h"
 
-double schorr_euchner(const int dim, GS_Info *gs_info, double R) {
+double schorr_euchner(const int dim, GS_Info *gs_info, long double R) {
     double *p = calloc(dim + 1, sizeof(double));
     double *v = calloc(dim, sizeof(double));
     double *c = calloc(dim, sizeof(double));
@@ -22,7 +22,7 @@ double schorr_euchner(const int dim, GS_Info *gs_info, double R) {
     }
     v[0] = 1;
 
-    double R_2 = R * R;
+    long double R_2 = R * R;
     int k = 0;
     int last_non_zero = 0;
